filter: make the MAX_VELOCITY_DIFF a double
The value is 0.001 which happens to be 0 when you cast it to int.

Introduced in e3554f38f8016ad89ffb7d9569eaa48769d48255

+++ src/filter.c
@@ -206,7 +206,7 @@ trackers_velocity_after_timeout(struct pointer_tracker *tracker,
 double
 trackers_velocity(struct pointer_trackers *trackers, uint64_t time)
 {
-	const int MAX_VELOCITY_DIFF = v_ms2us(1); /* units/us */
+	const double MAX_VELOCITY_DIFF = v_ms2us(1); /* units/us */
 	struct pointer_tracker *tracker;
 	double velocity;
 	double result = 0.0;
